#51 Fix dependency cycle
